Weigh the base texture more for textures created from blocks, since those usually have quite a bit of noise

+++ src/main/java/tconstruct/test/AbstractColoredTexture.java
@@ -174,6 +174,6 @@ public abstract class AbstractColoredTexture extends TextureAtlasSprite {
   }
 
   protected static int mult(int c1, int c2) {
-    return (int)(c1 * (c2/255.0));
+    return (int)((float)c1 * (c2/255f));
   }
 }
